Fix set display state after opening lid
When the underlying display device changes, we reset the screen state so that we can apply any new state to the new display device. To that end, we reset the screen state first, however we also need to reset the photonic modulator, otherwise if the new state is the same as the old one, photonic modulator won't set it because it'll think the state hasn't changed.

DisplayManagerService shouldn't ignore the new display state if it's the same as the old one, because the underlying display device might've changed and we still need to send the new state to the new device.

Bug: 442587030
Test: DisplayManagerServiceTest
Flag: com.android.server.display.feature.flags.fix_set_display_state_after_device_change
Change-Id: Id2bae8eee3fe208127c336cc38ce4f24a2fce79d