Reorder target display to top after cross-display task move
Without this change moving a focused task between displays (with
AppTask#moveTaskTo) does not change the order of DisplayContents under
RootWindowContainer which makes ATMS#getTopDisplayFocusedRootTask
consider only tasks on the source display.

The issue with such behavior is that:
  (1) any subsequent Activity#requestFullscreenMode calls on activities in that task fail unless the task is explicitly moved to front by the app after being moved across displays;
  (2) I believe users would not expect that a task will lose focus as a result of a programmatic movement across displays.

This change also makes an unfocused task gain focus as a result of
cross-display movement which, I believe, should be the expected behavior
– if "something" new appears on a display (a task in this case), the
focus pointer should move to that "something".

Fix: 436512189
Flag: EXEMPT BUGFIX
Test: WMShellUnitTests:DesktopTasksControllerTest
Change-Id: I9cd516241404739862659ce524d59d5765eed4ac