BouncyBall: Count all frame drops for app
Previously, in the trace metrics, we were only looking for
frame drops due to the app missing its deadline
(missed_app_frames).  But this app is very simple, and we
don't care just about it, but about the device as a whole.

Thus, we also want to include frame drops due to SurfaceFlinger
(missed_sf_frames), since this represents a device not providing
enough cycles to SF for a rather simple load.

We continue to ignore Buffer Stuffing issues, though, at least
in our initial version of our test setup.

Note this does not change the app or perfetto config at all, just
our metrics config.

We rename "missed_app_frames_bouncyball" to
"missed_frames_bouncyball".  This requires any automated setups
to change their logic as well, but is worth it to be more clear
about what we're noting.

Bug: 447197023
Test: "atest --include-preview-tests GtsPerformanceHostTestCases" still generally passes on our test devices
Test: "trace_processor_shell --summary --summary-metrics-v2 all --summary-spec trace_metrics_v2_spec.pbtx" on a trace with missed_sf_frames now counts these
Test: "trace_processor_shell --summary --summary-metrics-v2 all --summary-spec trace_metrics_v2_spec.pbtx" on a trace with missed_app_frames continues to count it
Flag: EXEMPT TEST_ONLY
Change-Id: I8d5ffec9e0c58ba9a9733520a205355b2a1f142c