BouncyBall: Allow up to 0.5 seconds for app focus
We've seen low end devices where the app fails to get focus within
0.1 seconds of starting to draw.  That triggers an immediate test
failure.

For now, we allow up to 0.5 seconds for the app to start drawing.
However, on devices where the app already has focus, we still start
measuring after 0.1 seconds, so we don't miss CPUs/GPUs clocking
down to save power.

This complicates the logic a bit more.  We rename a few variables
to further embrace the "relevant" frame nomenclature.

We update the comment in our perfetto config.  But, despite
increasing the potential app runtime by up to 0.4 seconds, we
keep the perfetto timeout the same since there was already a
decent amount of wiggle room with it.

Bug: 444496029
Test: "atest --include-preview-tests GtsPerformanceHostTestCases" on a upper-mid tier device; confirmed we're still starting after 0.1 seconds
Test: "atest --include-preview-tests GtsPerformanceHostTestCases" on a low-tier device; no longer immediately quits and runs the test
Flag: EXEMPT TEST_ONLY
Change-Id: I6d7a16b006e81c8ce5e35b8f0615588a6b4c25f9