jni: Migrate users of SurfaceTexture class behind flag.
This change migrates users of the SurfaceTexture class to a new
implementation behind the wb_surfacetexture feature flag. The new
implementation is part of a larger effort to refactor the graphics
buffer producer/consumer interfaces.

The key changes in this commit are:
- Introduction of a SurfaceTextureType typedef that resolves to either
  SurfaceTexture or LegacySurfaceTexture based on the feature flag.
- Use of build flag in JNI layer to switch between the old and new
  SurfaceTexture APIs.
- Updates to Camera, EGL, Surface, and TextureView to use the new
  SurfaceTexture API when the feature flag is enabled.

This migration is necessary to allow for the incremental rollout of the
new graphics buffer producer/consumer interfaces.

BYPASS_IGBP_IGBC_API_REASON=wb

Bug: 447613760
Flag: com.android.graphics.libgui.flags.wb_surfacetexture
Test: builds
Change-Id: I80453fed2c687b6d9416d68d5f69244377e7bbc6