Skip falsing when overlay is showing over communal
Previously when scene container is enabled, falsing is sometimes not
skipped when hub is showing, leading to unexpected behaviors. This is
because the falsing manager is looking for isIdleOnCommunal, and when an
overlay like bouncer or shade is transitioning on top of the hub, this
value returns false.

This change introduces a new flow that accounts for overlay transitions.

Test: atest CommunalSceneInteractorTest
Test: verified tapping widgets to launch activity does not go to
      lockscreen
Fix: 448198513
Flag: com.android.systemui.scene_container
Change-Id: I170fce8769da6b43b75e74f6d35ccfda8952e96a