[bouncer][desktop] Reveal password icon button on PasswordBouncer
Introduce an "eye" icon button in the PasswordBouncer text field that
allows the user to toggle the visibility of the entered password.

The button is only shown if the flag (see below) is enabled and the
config improveLargeScreenInteractionOnLockscreen is active.

This button is on the end of the text box. If the IME switcher icon
button is shown, that icon button is shown "before" the new button.

See https://b.corp.google.com/issues/427681136#comment16 for screencasts.

In a follow-up, additional time-based behaviors will be introduced:
- If the password has been revealed, it should be hidden again after 5s
  of no user interaction with the password input field
- If the "reveal password button" is available and the user is not
  interacting with the password input text field for 30s, the password
  should be reset.

Test: atest PasswordBouncerViewModelTest
Test: manual - button visible and usable on PasswordBouncer with corresponding flag and config
Flag: com.android.systemui.more_indicators_and_buttons_on_password_bouncer
Bug: 427681136
Change-Id: Iaab9bf807b744f70fab35bb0f50b897007ccd21a