Skip assigning ImeContainer layer in finish transaction
Since Id4c1c49ecb58b9f19718cb926f1ff2d76d791ad7, assignLayer
will always set to transaction for building finish transaction
of transition. That may cause potential transaction order issue,
such as the new layer is set when the transition is playing,
but the finish transaction sets to old layer when the transition
is finished.

Instead of skipping assigning layer of ImeContainer when a
transition is playing or collecting, just simply skip when
building finish transaction. Because it is common that IME
can show or hide when a transition is playing, it is better
to reflect current state immediately to avoid weird jumpcut
of IME appearance.

Bug: 443345373
Flag: EXEMPT BUGFIX
Test: WindowContainerTests#testAssignLayer
Change-Id: Icb50c80bc256e338d8fe197bebeb6fd2435f7eb2