Refactor DisplayShape to use lazy initialization and improved caching.
The `DisplayShape` class now defers the parsing of the shape specification string until it's actually needed, such as when `getPath()` or `toString()` is called. This is done by storing the parameters required to generate the spec string and using a new `initializeIfNeeded()` method.
Furthermore, we never actually store the path spec bytes in memory, even in the interim state.

The caching mechanism is enhanced to cache `DisplayShape` instances based on the source of creation: from a spec string, from resources, or as a default shape. This reduces redundant object creation.
In a typical app startup, I now see exactly 1 DisplayShape instance, where before there were 5 instances.

The `Parcelable` implementation is updated to write and read all necessary parameters to reconstruct the `DisplayShape` instance, leveraging the new caching and lazy initialization.
This saves us the trouble of having to copy the entire spec string (sometimes upwards of 2k characters) once per (duplicate!) instance of DisplayShape, and twice (once into the parcel, a second time out of the parcel). Instead we just copy a few integers that allow the recipient to construct the specified display shape.

This saves several expensive operations on app startup and ~22kb in unnecessary allocations in every app!
See measurements in: b/445026257#comment15

Bug: 445026257
Flag: EXEMPT BUGFIX
Change-Id: I14e963faf9491f0d5591d4759c41fd4d5c49c5d6