Skip tiny-framework-dump-test on JDK 25
Now we check certain build flags (currently only
RELEASE_BUILD_WITH_JDK_25), and if they are different from the expected
values, then we mark the test as pass.

Now this means when JDK 25 is enabled, tiny-framework-dump-test will
always pass.

When we introduce a flag for "-target 25", we should start using the flag too.

Bug: 395891737
Test: atest tiny-framework-dump-test
Test: Change the content of golden-output/00-hoststubgen-build-flags.txt
    and some other golden files, and make sure atest tiny-framework-dump-test passes.
Test: Change the content of some golden files,
    EXCEPT for golden-output/00-hoststubgen-build-flags.txt,
    and make sure atest tiny-framework-dump-test fails.
Flag: TEST_ONLY

Change-Id: Ia33ef3c1cb7940240c2a8c57411982b4eace380f