Use relayout result config/seqId as much as possible
This can often save a frame by using the return
value of relayoutWindow to pass-back the newest config
and seqId. This can technically be wrong, but that's
extremely rare and we were already basically doing
this before the always_seq_id flag anyways (so it isn't
a regression).

Bug: 439963876
Test: observe metrics
Flag: com.android.window.flags.always_seq_id_layout
Change-Id: I51f843ce2953168fb671c8186401210cebf64478