Prevent Growth Unlock  broadcast sent in OOBE
The Device Enter Directly signal has been detected on OOBE welcome screen
which lead to Growth App to start the syncing flow but fail to
receive the campaign. This change prevent the Device Enter Directly
broadcast to be sent during OOBE and make the first sync happens when
the user logs in session.

Bug: b/447623581
Test: GrowthInteractorTest
Flag: com.android.systemui.enable_desktop_growth
Change-Id: Iad4d2cf7126dd8140eedb76711f5069e0fc8fede