Create new install-time PPD flag.
Originally this flag was read-only but there is no reason for this to be
read-only. Creating a new read-write flag for this and deleting the
original one. This should be safe since we haven't started ramping the
original flag.

Note that we cannot just change the configuration of the original flag
since Gantry workflow has already been created.

Bug: 417788696
Test: presubmit
Flag: android.permission.flags.ppd_install_time_enabled
Change-Id: Ib78b8b2a9a2d1f6e010615886c0e89abeaee9b4e