Capture mProviderInfo in local variable before use
mProviderInfo is a volatile field, which indicates it may be used from
different threads. That means that null checks must happen on an
immutable snapshot of the field, in order to guarantee null dereferences.

For example:

if (mProviderInfo != null) mProviderInfo.doSomething();

should instead be:

var providerInfo = mProviderInfo;
if (providerInfo != null) providerInfo.doSomething();

So as to guarantee that a second thread doesn't mutate mProviderInfo
between the null check and the dereference.

Bug: b/445810711
Test: build
Flag: EXEMPT BUGFIX
Change-Id: Iccfa34a65d9ac5eb5b6c7e0f54ec1ddf6e2594bc