Skip full backup for packages updated during the backup process.
Store the `lastUpdateTime` of packages at the start of the full backup task. Before initiating the backup for each package, check if its `lastUpdateTime` has increased. If it has, the package was updated during the backup run, so skip it to avoid backing up potentially inconsistent data.

Change-Id: I901f644cd8be1a252cb41f0bb4d93cb319fcb5d4
Fixes: 377657814
Flag: EXEMPT BUGFIX