[PL] Send always-on-top change from the Core
Pinned layer's main property is to be always on top of other windows on
the current window. Since this is a FREEFORM window, but not a desktop
window, it should be removed from the desk.

Current implementation of the FreeformTaskTransitionObserver notifies
decoration view model and DesktopTaskChangeListener based on the
TransitionInfo.Change[]. Since WCT#setAlwayOnTop do not populate
ChangeInfo, those observers only get a Change containing a Desk because
Transtion#collectOrderindChanges collects tasks that have changed the
top order.

Current CL proposes to send a Change that indicates window entering AoT
state. It's assumed that exit notification is not required since
Transition#collectOrderChanges should catch up in case AoT task is
returned to the tree.

Bug: 399376001
Bug: 448091682
Flag: com.android.window.flags.enable_interactive_picture_in_picture
Test: atest WmTests:TransitionTests
Change-Id: Ieecaad067e8f54a4a2da97cd968f1ef7a57842b7