Implement an integration test for native processes
This commit introduces a simple integration test to check if a sample
native service is working as expected. This commit is basically a port
of ag/35284798. We may promote it to a CTS once the NDK APIs are made
public.

Bug: 427396671
Test: `atest NativeZygoteIntegrationTest` runs and passes when flags
are enabled and does not run when disabled
Flag: android.os.native_framework_prototype

Change-Id: I9d8fa41c37e5e6e777399e643c05c602f9a9a699