Write output directly to a file in systemfeatures-gen-tool
The Java runtime currently has a bug where sometimes it will output
a warning to stdout when running a java/kotlin tool. Since
systemfeatures-gen-tool produces output to stdout, that output gets
mixed with the warning and causes the build to fail. Write directly
to a file instead as a workaround.

Flag: EXEMPT PURE_REFACTOR
Bug: 449768258
Test: m systemfeatures-gen-srcs
Change-Id: I3b8371451871e5c58e3fd4d82ffbd52811d2ada4