a11y: Enable magnification UI with full screen only capability for mouse interactions
This change makes it possible for the mouse users to leverage the
magnification mode switch button and the setting panel with fullscreen
only capability. To avoid disrupting the experience for the touch users,
the UI remain invisible for touch interactions.

The feature is gated by a flag. When the flag is off, the bevahior
remains unchanged.

Key changes include:
- Added the `enable_magnification_ui_for_fullscreen` feature flag.
- Updated `MagnificationController` to track mouse interactions and show
  the UI accordingly when the flag is enabled.
- Modified `AccessibilityManagerService` to prevent the magnification
  button from being hidden in fullscreen-only mode.
- Fixed a bug where the capability of user state and the magnification
  controller might be out of sync.
- Refactored `WindowMagnificationSettings` to dynamically adjust the
  visibility of UI components based on the current magnification
  capability and mode.

Bug: 363166354
Flag: com.android.server.accessibility.enable_magnification_ui_for_fullscreen_only_capability
Test: atest SystemUITests:WindowMagnificationSettingsTest
Test: atest FrameworksServicesTests:AccessibilityManagerServiceTest
Test: atest FrameworksServicesTests:MagnificationControllerTest
Change-Id: Id08d44373d15f70a43248ae295c8f045f93d26a5