am: Prevent NPE when handling crash for unknown process
The `handleApplicationCrashInner` method can be called with a null
`ProcessRecord` if the crashing process has already been cleaned up by
the time its crash is processed. This could happen, for example, with
native crashes.

Previously, the method would unconditionally access the `ProcessRecord`
to trigger profiling, leading to a NullPointerException in the
system_server.

This CL adds a null check before accessing the `ProcessRecord` and adds
a `@Nullable` annotation to the parameter to prevent the crash and make
the nullability explicit.

Bug: 449277211
Test: m services.core
Flag: android.os.profiling.profiling_trigger_oom

Change-Id: I02ec9a858b0cdb42af52d4f2c886fe8ec97eafb7