a11y: Add an accessibility button on bouncer
This change adds an accessibility button to the bottom right corner of
the bouncer screen. This is controlled by the aconfig flag below as well
as a config value that will only be enabled on specific form factors,
e.g. desktop.

UX, design, and screenshots are included in the linked bug.

Bug: 449824070
Test: atest SystemUITests
Flag: com.android.systemui.bouncer_accessibility_button_for_desktop
Change-Id: Iaaa8f69f9a4010d9706e983cb5590ace3250fa76