Policy streamlining: Support storing policies in DPE from DPMS.setPolicy
This requires access to the `PolicyDefinition<T>` as it is used in the
DevicePolicyEngine to identify the policy.
For now I force every `PolicyHandler` to take this policy definition as
a constructor argument, however longer term we will want to change it so
people no longer have to hand-write this policy definition.

Test: atest FrameworksServicesTests:com.android.server.devicepolicy.handlers.PolicyHandlerTest
Test: btest "android.devicepolicy.cts.ScreenCaptureDisabledTest"
Bug: 444415053
Bug: 447152592
Flag: android.app.admin.flags.policy_streamlining
Change-Id: I6aa31cfee0e69070aa5c2395d83d0bca4130eefc