Desktop screenshots: Delay fullscreen screenshot to hide UI
Sometimes the pre-capture UI will be captured in a fullscreen
screenshot. Although rare, it appears to happen because the UI does not
have enough time to hide before the screenshot is taken. Add a 100ms
delay which should provide enough time to dismiss the UI.

Bug: 435225255
Test: atest PreCaptureViewModelTest
Flag: com.android.systemui.large_screen_screencapture
Change-Id: If5d3bf0581832383805454a245116d3593b6c638