Deprecate DesktopRepository#addTask in favor of DesktopRepository#addTaskToDesk
Corrects an assumption made by addTask that any task added to the repository should always go to the active desk. Instead, try to find a desk the task belongs to first. Failing that, fall back to the active one. This fixes a bug where restoring multiple desks simultaneously to a display can result in all tasks being forced into the active desk.

To access the appropriate desk, establishes a dependency on DesksOrganizer in DesktopTaskChangeListener and changes the vice versa to a listener.

Bug: 440248564
Test: Manual; disconnect/reconnect with 10 tasks open in different desks; verify they stay separate
Flag: com.android.window.flags.enable_display_reconnect_interaction
Change-Id: I12ca0de9265ccd7b7a4c840513ee734fe40390f5