Clear windowing mode on reparentTasks
This CL ensures that the windowing mode of a task which is going to be
reparented to the split root is cleared.
On freeform displays, a fullscreen task has the override fullscreen mode
rather than W_M_UNDEFIEND. Therefore its override windowing mode must be
cleared, allowing the leaf task to inherit the windowing mode of the
split root.

Flag: com.android.window.flags.enable_desktop_first_launch_adjacent_bugfix
Bug: 440254150
Bug: 440299242
Test: WindowContainerTransactionTests
Change-Id: I7391c99bb20a53cabd16bf912a46c2d0b2e363f4