Introduce HoistedSceneTransitionLayoutState
This CL introduces a new HoistedSceneTransitionLayoutState interface and
factory that better separates the mutation operations that require to
be bound to a UI (because we need SceneTransitions definitions and/or
MotionScheme) from those that don't. This allows to have a more
self-contained HoistedSTLState interface that does not require to pass
the SceneTransitions/MotionScheme when creating it, or setting them to a
placeholder no-op value to then update them later during composition.

This CL should also help get rid of the anti-pattern used by
SceneContainer [1], Communal [2] and BundleHeader [3] where the STLState
is created during composition and then forwarded back to the data layer.

This new HoistedSTLState interface now exposes a
`uiBound: MutableSTLState?` that returns a non-null state interface of
all the mutation operations that require a
UI/SceneTransitions/MotionScheme.

I wrote this CL on purpose so that the existing MutableSTLState API does
not change and that existing code should still compile and have the same
behavior as before.

[1] http://shortn/_Ny3EXhJNiu
[2] http://shortn/_LTUUvgrxYk
[3] http://shortn/_woX0GOEroW

Bug: 450236706
Bug: 449675581
Test: atest PlatformComposeSceneTransitionLayoutTests
Flag: NONE syntax refactoring + new unused class
Change-Id: I3fce1c898d6c15ac10c123403519dfeee6c6aa54