New API method to launch a PendingIntent in a bubble
New method is launchInBubble. It will take a PendingIntent argument.
The PendingIntent must be for activity launch.
This new PendingIntent argument is launched in a bubble without
modifications. Allowing callers to specify the intent flags as they see
fit.
Also ensures that the uid of the caller is used when launching the
intent.

Bug: 407149510
Test: use the test app to launch a bubble, check that the caller id is
  set to the test app
Flag: com.android.window.flags.enable_experimental_bubbles_controller
Change-Id: Ie200855f1062f036793bfd4c6d909417e3b81297