Update clip data dump in BubbleExpandedView
Print out individual clip values instead of a clip rect. These values
indicate the amount to clip from each side, left, right etc. If all the
values are 0, it means there is no clipping. If we print it out as a
mClipRect(0,0,0,0), it can be confused with an actual clip rect. And in
this case would mean the whole view is clipped. As clipRect would result
in an empty rectangle.

Bug: na
Test: na
Flag: EXEMPT DEBUG
Change-Id: I41f9742451e46db4beb93e4918d0088113725f65