Update win reporting with remote insets control
When system bars are controlled with a remote insets controller, ensure
that the new window can be controlled by the remote controller before
not requiring the app to cover the entire display. Also update the top
fullscreen app with any app with fullscreen attrs even if it doesn't
cover the entire screen when controlled remotely.

Bug: 443862689
Test: atest InsetsPolicyTest
Flag: NONE bugfix
Change-Id: Ic204d2016026460321da6a58088cc2f5cbb0aa17