am: Remove push_activity_state_to_oomadjuster flag
The `push_activity_state_to_oomadjuster` flag was used to migrate the
OomAdjuster from a pull-based model for activity-related state to a
more efficient push-based model.

Now that the migration is complete and the feature has been enabled.
This CL removes the flag and all associated legacy code paths. This
simplifies the codebase by making the push-based logic the only
implementation for retrieving activity state within the OomAdjuster.

Bug: 401350380
Bug: 441408003
Fix: 446551688
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
            ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT FLAG_REMOVAL

Change-Id: I00c7fda4896836846606442c8eb3fd91f62fa120