Temporary fix for unresponsive password bouncer
This adds a temporary workaround to fix the issue which causes the
password bouncer text field to become unresponsive after sleep.
For the temporary fix, the focus will be reset after the scene
transition to bouncer is finished. This will cause the text field to
accept input again without recomposing the scene or blocking input
during the scene transition.

Flag: com.android.systemui.scene_container
Bug: 445347474
Bug: 451546441
Test: manual unlocking on brya and cf tablet after sleep
Test: atest BouncerContentTest, BouncerContentComposeTest
Change-Id: I9340d4c7bb756c643bf8dc09cabc56e7a5c570fa