ComputerControl: Do not leak VirtualDisplay token to client
The VirtualDisplay for the ComputerControlSession is owned by the
system. Instead of passing the token to the client for it to initialize
the session with a Surface, have the client pass its Surface to VDM.

This way, VDM is fully in control of the lifecycle of the session. The
client does not have access the VD's token so it cannot update the
VD's surface by itself.

This is a pure refactor.

Bug: 442540670
Test: Presubmit
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: I6256ba7c8f59d7fe8ae6e61b250f0cc2b38bcb06