[2/N] Implement first-run prompt prototype
This CL implements a prototype of the first-run prompt.
The prompt is shown when a new app window is opened if the app is opened
due to the app link capture for the first time.
Please note that this is a prototype and UI/trigger condition details
will be adjusted later.

Flag: com.android.window.flags.enable_enhanced_app_to_web_transition
Bug: 451760697
Test: manual
Change-Id: Ib37698b08859cac2b9cf459500ccb191ccf0bdcc