[bouncer][desktop] Show sign-in button for PasswordBouncer
For PasswordBouncer, show a "Sign in" button on the bottom of
BouncerContent.
This button is only enabled if there is some text in PasswordBouncer's
TextFieldState ([readyToAuthenticate] StateFlow).

Whether the button is displayed is controlled by
[AuthMethodBouncerViewModel.showSignInButton], which is `false` by default
and PasswordBouncerViewModel overrides it to `true`.

PinBouncer and PatternBouncer don't get the button, but still offer a
[readyToAuthenticate] StateFlow for consistency.

See b/427681136#comment28 for screenshots.

Test: atest PasswordBouncerViewModelTest BouncerOverlayContentViewModelTest
Test: manual - look at password bouncer on "desktop" with Flag on :-)
Flag: com.android.systemui.more_indicators_and_buttons_on_password_bouncer
Bug: 427681136
Change-Id: Ibc106bb8efc1e6927e277601547067d8276d5fcc