Ignore keyguardFadingAway when scenes are enabled
The value of keyguardFadingAway can sometimes get stuck on true when
scenes are enabled, which causes the shade window to remain visible
and eat touches. keyguardFadingAway is no longer an imporant value,
and the team has wanted to remove it for a long time. We believe that
scenes make it obsolete and we can flag it off safely, which will
also fix b/449982657.

Fixes: 449982657
Test: added a test and ran it with both flag values
Flag: com.android.systemui.scene_container
Change-Id: Iaf52662d41c5ac673ec59820048778578367c979