Close And Reopen ScreencaptureUI For File Picker
Originally, the file picker was launched behind the screencapture UI.
This caused the toolbar to be shown above the file picker & also made
the file picker uninteractable unless the toolbar was closed. In this
CL, we will temporarily hide the screencapture UI (make it invisible and
you can touch through it), and then bring it back once the user is done
selecting their files. This method also allows for the ui states (eg
partial screenshot region box, settings, etc.) to be maintained when the
screencapture UI is brought back up.

Bug: 444277547
Test: Manual
Flag: com.android.systemui.large_screen_screenshot_save_location
Change-Id: I9e9829b62f542c9f58bf78cdf151f1042b6b8577