Changed ActivityStarter so it checks allowlisted activities per user.
Previously, it was only explicitly checking if the user was the HSU
(Headless System User) and calling UMI.getActiviesAllowlist(type),
but now it calls UMI.getActiviesAllowlist(userId) for all users.

But since that call happens in a critical path, UMS still mantains
a map of allowlist per user type, and caches the allowlist per user
as users start (and stop).

Test: atest FrameworksMockingServicesTests:UserActivitiesAllowlistTest WmTests:ActivityStarterHsuAllowlistIntegrationTests
Test: adb shell dumpsys user --activities-allowlist
Bug: 412177078
Flag: android.multiuser.hsu_allowlist_activities

Change-Id: I15625b5671b3ef23f3de621fd98aaa8562f84323