OemLockService.setOemUnlockAllowed allows SYSTEM not just Admin
setOemUnlockAllowed...() methods check that the caller is an Admin user.
Historically, the System user has always been considered an Admin user,
so this always included user 0. But on some devices, user 0 will not be
an Admin user; for historical reasons, we may need to allow user 0 to
call this method, even though it isn't an Admin.

We therefore relax the requirement slightly so as to keep the current
status quo.

Note that there are already separate requirements that the caller has
MANAGE_CARRIER_OEM_UNLOCK_STATE or MANAGE_USER_OEM_UNLOCK_STATE, so this
relaxation is not large.

This parallels change I4a017421c846390670aea5f95a75694eb586aa07 for
PersistentDataBlockService.

Bug: 443134869
Test: OemLockServiceTest
Flag: android.multiuser.hsu_not_admin
Change-Id: I1bde6d16e7294a2a0c456f6ce8550032603836c7