Continue sending callbacks to non rr subsribers when in background
We want to not send refresh rate changed callbacks when the client is in the background. However the current implementation drops out all the events. This bug was recently introduced when we bacthed the events to an eventmask approach

Bug: 449949226
Test: Manual
Flag: com.android.server.display.feature.flags.send_non_rr_callbacks_when_in_background
Change-Id: Iea4a3e083ac7ed869ed9b6ed08b520238a08d5c4