ComputerControlSession: Introduce "unknown" constants for callbacks
Add "unknown" contstants for APIs with enums that the system reports to
the client. This needs to be established to handle the case where the
system reports a new enum value that the client library does not yet
know about, in which case the library can convert the new constant to
the unknown enum, until the client library is updated to handle the new
constant.

Bug: 442540670
Test: Presubmit
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: Idecc1617ff384b2e480864ad93f66c1cb7bfc116