Add an approximate cache for isRecentsComponentHomeActivity
Currently it is only called for dumping. The invocation of
getDefaultHomeActivity may be heavy if the frequency of dump is high.

The cache is unset when resuming a different home or switching user.
The result may be inaccurate if home is not launched after calling
setDefaultApplication to change default home. But maybe there is
no such test which needs to verify the case so far.

Though an OnRoleHoldersChangedListener will be accurate, but an extra
registration only for debugging seems to be also wasteful.

Bug: 112837428
Flag: EXEMPT TEST_ONLY
Test: FlickerTests
Change-Id: I31e827eb7c05f84c427d3729fc95bbba6add255d