Use BackupManagerMonitor from KeyValueBackupTask constructor if it's not null.
We should always initialize the reporter with the provided monitor if it's not null. The monitor will be obtained from the transport once the flag in GMS Core that prevents us from sending the monitor in the params is ramped up.

Change-Id: Id8eec2fb1ccb843d3c513c5ddf90dbc0ea9bdc35
Bug: 443738546
Bug: 380051012
Bug: 450630210
Flag: com.android.server.backup.enable_kv_backup_logs_from_transport_with_proper_flow_id
Tested: CtsBackupTestCases and CtsBackupHostTestCases